chore: more cleanups
