+++ src/screens/conversations/components/conversation-sheet/StatusListSheet.tsx
@@ -9,9 +9,10 @@ import { TickIcon } from '@/svg-icons';
 import { tailwind } from '@/theme';
 import { StatusCollection } from '@/types';
 import { getStatusTypeIcon, useHaptic } from '@/utils';
-import { BottomSheetHeader, Icon, StatusOptions } from '@/components-next';
+import { BottomSheetHeader, Icon } from '@/components-next';
 import { useAppDispatch, useAppSelector } from '@/hooks';
 import i18n from '@/i18n';
+import { StatusOptions } from '@/types';
 
 type StatusCellProps = {
   value: StatusCollection;
